src/requirements.txt
cffi==1.17.1
numpy==2.2.2
pedalboard==0.9.16
pycparser==2.22
python-dotenv==1.0.1
soundfile==0.13.1
